backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-dotenv==1.1.1
apscheduler==3.11.0
httpx==0.28.1
beautifulsoup4==4.13.4
rapidfuzz==3.14.1
python-slugify==8.0.4
email-validator==2.2.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
ruff==0.12.10
